split ControlMass to a few subcomponents and move them to control.py
