+++ aviary/subsystems/mass/gasp_based/fixed.py
@@ -3,6 +3,7 @@ import openmdao.api as om
 from openmdao.components.ks_comp import KSfunction
 
 from aviary.constants import GRAV_ENGLISH_LBM, RHO_SEA_LEVEL_ENGLISH
+from aviary.subsystems.mass.gasp_based.control import ControlMassGroup
 from aviary.subsystems.mass.gasp_based.engine import EngineMassGroup
 from aviary.utils.math import dSigmoidXdx, sigmoidX
 from aviary.variable_info.enums import FlapType
@@ -2163,289 +2164,6 @@ class HighLiftMass(om.ExplicitComponent):
         )
 
 
-class ControlMass(om.ExplicitComponent):
-    """
-    Computation of total mass of cockpit controls, fixed wing controls, and SAS,
-    and mass of surface controls.
-    """
-
-    def setup(self):
-        add_aviary_input(self, Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT, units='unitless')
-        add_aviary_input(self, Aircraft.Wing.AREA, units='ft**2')
-        add_aviary_input(self, Aircraft.Design.GROSS_MASS, units='lbm')
-        add_aviary_input(self, Aircraft.Wing.ULTIMATE_LOAD_FACTOR, units='unitless')
-        self.add_input('min_dive_vel', val=700, units='kn', desc='VDMIN: dive velocity')
-        add_aviary_input(self, Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT, units='unitless')
-        add_aviary_input(
-            self, Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS, units='lbm'
-        )
-        add_aviary_input(self, Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER, units='unitless')
-        add_aviary_input(self, Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER, units='unitless')
-        add_aviary_input(
-            self,
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER,
-            units='unitless',
-        )
-        add_aviary_input(self, Aircraft.Controls.CONTROL_MASS_INCREMENT, units='lbm')
-
-        add_aviary_output(self, Aircraft.Controls.COCKPIT_CONTROL_MASS, units='lbm')
-        add_aviary_output(self, Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS, units='lbm')
-        add_aviary_output(self, Aircraft.Controls.MASS, units='lbm')
-        add_aviary_output(self, Aircraft.Wing.SURFACE_CONTROL_MASS, units='lbm')
-
-    def setup_partials(self):
-        self.declare_partials(Aircraft.Controls.MASS, '*')
-        self.declare_partials(
-            Aircraft.Controls.COCKPIT_CONTROL_MASS,
-            [
-                Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER,
-                Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT,
-                Aircraft.Design.GROSS_MASS,
-            ],
-        )
-        self.declare_partials(
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS,
-            [
-                Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER,
-                Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS,
-            ],
-        )
-        self.declare_partials(
-            Aircraft.Wing.SURFACE_CONTROL_MASS,
-            [
-                Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT,
-                Aircraft.Wing.AREA,
-                Aircraft.Design.GROSS_MASS,
-                Aircraft.Wing.ULTIMATE_LOAD_FACTOR,
-                'min_dive_vel',
-                Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT,
-                Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER,
-                Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER,
-            ],
-        )
-
-    def compute(self, inputs, outputs):
-        c_mass_trend_wing_control = inputs[Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT]
-
-        wing_area = inputs[Aircraft.Wing.AREA]
-        gross_wt_initial = inputs[Aircraft.Design.GROSS_MASS] * GRAV_ENGLISH_LBM
-        ULF = inputs[Aircraft.Wing.ULTIMATE_LOAD_FACTOR]
-
-        c_mass_trend_cockpit_control = inputs[Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT]
-
-        stab_aug_base_wt = (
-            inputs[Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS] * GRAV_ENGLISH_LBM
-        )
-        CK15 = inputs[Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER]
-        CK18 = inputs[Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER]
-        CK19 = inputs[Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER]
-        delta_control_wt = inputs[Aircraft.Controls.CONTROL_MASS_INCREMENT] * GRAV_ENGLISH_LBM
-        min_dive_vel = inputs['min_dive_vel']
-
-        dive_param = (1.15 * min_dive_vel) ** 2 / 391.0
-
-        intermediate_control_wt = (
-            c_mass_trend_wing_control
-            * wing_area**0.317
-            * (gross_wt_initial / 1000.0) ** 0.602
-            * ULF**0.525
-            * dive_param**0.345
-        )
-        intermediate_cockpit_control_wt = (
-            c_mass_trend_cockpit_control * (gross_wt_initial / 1000.0) ** 0.41
-        )
-        cockpit_control_wt = CK15 * intermediate_cockpit_control_wt
-        wing_control_wt = CK18 * (intermediate_control_wt - intermediate_cockpit_control_wt)
-        stab_control_wt = CK19 * stab_aug_base_wt
-
-        outputs[Aircraft.Controls.COCKPIT_CONTROL_MASS] = cockpit_control_wt / GRAV_ENGLISH_LBM
-        outputs[Aircraft.Wing.SURFACE_CONTROL_MASS] = wing_control_wt / GRAV_ENGLISH_LBM
-        outputs[Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS] = (
-            stab_control_wt / GRAV_ENGLISH_LBM
-        )
-        outputs[Aircraft.Controls.MASS] = (
-            cockpit_control_wt + wing_control_wt + stab_control_wt + delta_control_wt
-        ) / GRAV_ENGLISH_LBM
-
-    def compute_partials(self, inputs, J):
-        c_mass_trend_wing_control = inputs[Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT]
-
-        wing_area = inputs[Aircraft.Wing.AREA]
-        gross_wt_initial = inputs[Aircraft.Design.GROSS_MASS] * GRAV_ENGLISH_LBM
-        ULF = inputs[Aircraft.Wing.ULTIMATE_LOAD_FACTOR]
-
-        c_mass_trend_cockpit_control = inputs[Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT]
-
-        stab_aug_base_wt = inputs[Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS]
-        CK15 = inputs[Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER]
-        CK18 = inputs[Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER]
-        CK19 = inputs[Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER]
-        min_dive_vel = inputs['min_dive_vel']
-
-        dive_param = (1.15 * min_dive_vel) ** 2 / 391.0
-
-        intermediate_control_wt = (
-            c_mass_trend_wing_control
-            * wing_area**0.317
-            * (gross_wt_initial / 1000.0) ** 0.602
-            * ULF**0.525
-            * dive_param**0.345
-        )
-
-        intermediate_cockpit_control_wt = (
-            c_mass_trend_cockpit_control * (gross_wt_initial / 1000.0) ** 0.41
-        )
-
-        cockpit_control_wt = CK15 * intermediate_cockpit_control_wt
-        wing_control_wt = CK18 * (intermediate_control_wt - intermediate_cockpit_control_wt)
-
-        dSCW_dSWCC = (
-            CK18
-            * wing_area**0.317
-            * (gross_wt_initial / 1000.0) ** 0.602
-            * ULF**0.525
-            * dive_param**0.345
-        )
-
-        dSCW_dWA = 0.317 * (
-            c_mass_trend_wing_control
-            * wing_area ** (0.317 - 1)
-            * (gross_wt_initial / 1000.0) ** 0.602
-            * ULF**0.525
-            * dive_param**0.345
-        )
-
-        dSCW_dWG = 0.602 * c_mass_trend_wing_control * wing_area**0.317 * (
-            gross_wt_initial / 1000.0
-        ) ** (0.602 - 1) * (
-            1 / 1000
-        ) * ULF**0.525 * dive_param**0.345 - 0.41 * c_mass_trend_cockpit_control * (
-            gross_wt_initial / 1000.0
-        ) ** (0.41 - 1) * (1 / 1000)
-
-        dSCW_dULF = (
-            0.525
-            * c_mass_trend_wing_control
-            * wing_area**0.317
-            * (gross_wt_initial / 1000.0) ** 0.602
-            * ULF ** (0.525 - 1)
-            * dive_param**0.345
-        )
-        dSCW_dMDV = (
-            0.345
-            * c_mass_trend_wing_control
-            * wing_area**0.317
-            * (gross_wt_initial / 1000.0) ** 0.602
-            * ULF**0.525
-            * dive_param ** (0.345 - 1)
-            * 2
-            * 1.15
-            * min_dive_vel
-            * 1.15
-            / 391
-        )
-        dSCW_dCCWC = -CK18 * ((gross_wt_initial / 1000.0) ** 0.41)
-
-        J[
-            Aircraft.Wing.SURFACE_CONTROL_MASS,
-            Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT,
-        ] = dSCW_dSWCC / GRAV_ENGLISH_LBM
-
-        J[Aircraft.Controls.MASS, Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT] = (
-            dSCW_dSWCC / GRAV_ENGLISH_LBM
-        )
-
-        J[Aircraft.Wing.SURFACE_CONTROL_MASS, Aircraft.Wing.AREA] = dSCW_dWA / GRAV_ENGLISH_LBM
-
-        J[Aircraft.Controls.MASS, Aircraft.Wing.AREA] = CK18 * dSCW_dWA / GRAV_ENGLISH_LBM
-
-        J[Aircraft.Wing.SURFACE_CONTROL_MASS, Aircraft.Design.GROSS_MASS] = CK18 * dSCW_dWG
-
-        J[Aircraft.Controls.MASS, Aircraft.Design.GROSS_MASS] = (
-            0.41
-            * CK15
-            * (
-                c_mass_trend_cockpit_control
-                * (gross_wt_initial / 1000.0) ** (0.41 - 1)
-                * (1 / 1000)
-            )
-            + (CK18 * dSCW_dWG)
-        ) / GRAV_ENGLISH_LBM
-
-        J[Aircraft.Wing.SURFACE_CONTROL_MASS, Aircraft.Wing.ULTIMATE_LOAD_FACTOR] = (
-            dSCW_dULF / GRAV_ENGLISH_LBM
-        )
-
-        J[Aircraft.Wing.SURFACE_CONTROL_MASS, Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER] = (
-            intermediate_control_wt - intermediate_cockpit_control_wt
-        ) / GRAV_ENGLISH_LBM
-
-        J[Aircraft.Controls.MASS, Aircraft.Wing.ULTIMATE_LOAD_FACTOR] = (
-            CK18 * dSCW_dULF / GRAV_ENGLISH_LBM
-        )
-        J[Aircraft.Wing.SURFACE_CONTROL_MASS, 'min_dive_vel'] = dSCW_dMDV / GRAV_ENGLISH_LBM
-        J[Aircraft.Controls.MASS, 'min_dive_vel'] = CK18 * dSCW_dMDV / GRAV_ENGLISH_LBM
-
-        J[
-            Aircraft.Wing.SURFACE_CONTROL_MASS,
-            Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT,
-        ] = dSCW_dCCWC / GRAV_ENGLISH_LBM
-
-        J[
-            Aircraft.Controls.MASS,
-            Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT,
-        ] = (CK15 * (gross_wt_initial / 1000.0) ** 0.41 + dSCW_dCCWC) / GRAV_ENGLISH_LBM
-
-        J[
-            Aircraft.Controls.MASS,
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS,
-        ] = CK19
-
-        J[Aircraft.Controls.MASS, Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER] = (
-            intermediate_cockpit_control_wt / GRAV_ENGLISH_LBM
-        )
-
-        J[Aircraft.Controls.MASS, Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER] = (
-            wing_control_wt / CK18 / GRAV_ENGLISH_LBM
-        )
-
-        J[
-            Aircraft.Controls.MASS,
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER,
-        ] = stab_aug_base_wt
-        J[Aircraft.Controls.MASS, Aircraft.Controls.CONTROL_MASS_INCREMENT] = 1
-
-        J[
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS,
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER,
-        ] = stab_aug_base_wt
-        J[
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS,
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS,
-        ] = CK19
-
-        J[
-            Aircraft.Controls.COCKPIT_CONTROL_MASS,
-            Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER,
-        ] = c_mass_trend_cockpit_control * (gross_wt_initial / 1000.0) ** 0.41 / GRAV_ENGLISH_LBM
-
-        J[
-            Aircraft.Controls.COCKPIT_CONTROL_MASS,
-            Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT,
-        ] = CK15 * (gross_wt_initial / 1000.0) ** 0.41 / GRAV_ENGLISH_LBM
-
-        J[
-            Aircraft.Controls.COCKPIT_CONTROL_MASS,
-            Aircraft.Design.GROSS_MASS,
-        ] = (
-            CK15
-            * c_mass_trend_cockpit_control
-            * 0.41
-            * (gross_wt_initial / 1000.0) ** -0.59
-            / 1000.0
-        )
-
-
 class TotalLandingGearMass(om.ExplicitComponent):
     """Computation of total mass of landing gear."""
 
@@ -2684,7 +2402,7 @@ class FixedMassGroup(om.Group):
 
         self.add_subsystem(
             'controls',
-            ControlMass(),
+            ControlMassGroup(),
             promotes_inputs=['*'],
             promotes_outputs=['*'],
         )
