+++ aviary/subsystems/mass/gasp_based/test/test_fixed.py
@@ -8,7 +8,6 @@ from openmdao.utils.testing_utils import use_tempdirs
 from aviary import constants
 from aviary.constants import RHO_SEA_LEVEL_ENGLISH
 from aviary.subsystems.mass.gasp_based.fixed import (
-    ControlMass,
     ElectricAugmentationMass,
     FixedMassGroup,
     HighLiftMass,
@@ -545,65 +544,6 @@ class HighLiftTestCase(unittest.TestCase):
         assert_check_partials(data, atol=5e-10, rtol=1e-12)
 
 
-# this is the large single aisle 1 V3 test case
-@use_tempdirs
-class ControlMassTestCase(unittest.TestCase):
-    def setUp(self):
-        self.prob = om.Problem()
-        self.prob.model.add_subsystem('control_mass', ControlMass(), promotes=['*'])
-
-        self.prob.model.set_input_defaults(
-            Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT, val=0.95, units='unitless'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Wing.AREA, val=1392.1, units='ft**2'
-        )  # bug fixed value
-        self.prob.model.set_input_defaults(
-            Aircraft.Design.GROSS_MASS, val=175400, units='lbm'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Wing.ULTIMATE_LOAD_FACTOR, val=3.951, units='unitless'
-        )  # bug fixed value
-        self.prob.model.set_input_defaults(
-            'min_dive_vel', val=420, units='kn'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT, val=16.5, units='unitless'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS, val=0, units='lbm'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER, val=1, units='unitless'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER, val=1, units='unitless'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER, val=1, units='unitless'
-        )  # bug fixed value and original value
-        self.prob.model.set_input_defaults(
-            Aircraft.Controls.CONTROL_MASS_INCREMENT, val=0, units='lbm'
-        )  # bug fixed value and original value
-
-        self.prob.setup(check=False, force_alloc_complex=True)
-
-    def test_case1(self):
-        self.prob.run_model()
-
-        expected_values = {
-            Aircraft.Controls.MASS: 3945,
-        }
-        tol = 5e-4
-
-        for var_name, expected_val in expected_values.items():
-            with self.subTest(var=var_name):
-                assert_near_equal(self.prob[var_name], expected_val, tol)
-
-        data = self.prob.check_partials(out_stream=None, method='cs')
-        assert_check_partials(data, atol=1e-11, rtol=1e-12)
-
-
 @use_tempdirs
 class GearTestCase1(unittest.TestCase):  # this is the large single aisle 1 V3 test case
     def setUp(self):
@@ -943,6 +883,9 @@ class FixedMassGroupTestCase1(unittest.TestCase):
             Aircraft.HorizontalTail.MASS: 2285,
             Aircraft.VerticalTail.MASS: 2312,
             Aircraft.Wing.HIGH_LIFT_MASS: 4082.1,
+            Aircraft.Controls.COCKPIT_CONTROL_MASS: 137.25749725,
+            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS: 0.0,
+            Aircraft.Wing.SURFACE_CONTROL_MASS: 3807.92115815,
             Aircraft.Controls.MASS: 3945,
             Aircraft.LandingGear.TOTAL_MASS: 7511,
             Aircraft.Propulsion.TOTAL_ENGINE_MASS: 12606,
@@ -1607,59 +1550,6 @@ class BWBHighLiftTestCase(unittest.TestCase):
 
         data = self.prob.check_partials(out_stream=None, method='cs', show_only_incorrect=True)
         assert_check_partials(data, atol=5e-10, rtol=1e-12)
-
-
-@use_tempdirs
-class BWBControlMassTestCase(unittest.TestCase):
-    """GAST BWB model"""
-
-    def setUp(self):
-        prob = self.prob = om.Problem()
-        prob.model.add_subsystem('control_mass', ControlMass(), promotes=['*'])
-
-        prob.model.set_input_defaults(
-            Aircraft.Wing.SURFACE_CONTROL_MASS_COEFFICIENT, 0.5, units='unitless'
-        )
-        prob.model.set_input_defaults(Aircraft.Wing.AREA, 2142.85714286, units='ft**2')
-        prob.model.set_input_defaults(Aircraft.Design.GROSS_MASS, 150000, units='lbm')
-        prob.model.set_input_defaults(
-            Aircraft.Wing.ULTIMATE_LOAD_FACTOR, 3.97744787, units='unitless'
-        )
-        prob.model.set_input_defaults('min_dive_vel', 420, units='kn')
-        prob.model.set_input_defaults(
-            Aircraft.Design.COCKPIT_CONTROL_MASS_COEFFICIENT, 16.5, units='unitless'
-        )
-        prob.model.set_input_defaults(
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_BASE_MASS, 0, units='lbm'
-        )
-        prob.model.set_input_defaults(
-            Aircraft.Controls.COCKPIT_CONTROL_MASS_SCALER, 1, units='unitless'
-        )
-        prob.model.set_input_defaults(
-            Aircraft.Wing.SURFACE_CONTROL_MASS_SCALER, 1, units='unitless'
-        )
-        prob.model.set_input_defaults(
-            Aircraft.Controls.STABILITY_AUGMENTATION_SYSTEM_MASS_SCALER, 1, units='unitless'
-        )
-        prob.model.set_input_defaults(Aircraft.Controls.CONTROL_MASS_INCREMENT, 0, units='lbm')
-
-        prob.setup(check=False, force_alloc_complex=True)
-
-    def test_case1(self):
-        self.prob.run_model()
-
-        expected_values = {
-            Aircraft.Wing.SURFACE_CONTROL_MASS: 2045.5556421,
-            Aircraft.Controls.MASS: 2174.28611375,
-        }
-        tol = 1e-7
-
-        for var_name, expected_val in expected_values.items():
-            with self.subTest(var=var_name):
-                assert_near_equal(self.prob[var_name], expected_val, tol)
-
-        data = self.prob.check_partials(out_stream=None, method='cs')
-        assert_check_partials(data, atol=1e-11, rtol=1e-12)
 
 
 class BWBGearTestCase(unittest.TestCase):
